build: split kz, forkzio from libzio
Move kz to src/modules/libkz and build between src/modules/kvs
(which it requires) and src/modules/wreck (which requires it).
Build as a noinst convenience library for now as kz_*() is not
currently exported as part of the public api.

Move forkzio.[ch] to src/cmd and link directly with flux-zio as
that is its only user.

dependencies and leaking symbols I think are resolved, re-arranging
kz helped a great deal, just had to ensure that it didn't include
any symbols that should not be leaked.

zio factored out of common, now a fluxlib
