+++ src/cmd/flux-zio.c
@@ -34,11 +34,13 @@
 #include <json.h>
 #include <flux/core.h>
 
+#include "forkzio.h"
+
 #include "src/common/libutil/log.h"
 #include "src/common/libutil/xzmalloc.h"
-#include "src/common/libutil/zio.h"
-#include "src/common/libzio/kz.h"
-#include "src/common/libzio/forkzio.h"
+#include "src/modules/libzio/zio.h"
+#include "src/modules/libkz/kz.h"
+#include "src/cmd/forkzio.h"
 
 
 typedef struct {
