+++ src/modules/libzio/flux-zio-test.c
@@ -32,7 +32,7 @@
 #include <json.h>
 #include <flux/core.h>
 
-#include "src/common/libutil/zio.h"
+#include "src/modules/libzio/zio.h"
 
 int output_thread_cb (flux_t f, void *zs, short revents, zio_t z)
 {
