+++ src/modules/libzio/zio.c
@@ -361,7 +361,7 @@ int zio_set_close_cb (zio_t zio, zio_close_f closef)
     return (0);
 }
 
-static int zio_fd_read (zio_t zio, void *dst, int len)
+static int zio_read (zio_t zio, void *dst, int len)
 {
     assert (zio != NULL);
     assert (zio->magic == ZIO_MAGIC);
@@ -473,7 +473,7 @@ int zio_flush (zio_t zio)
         zio_debug (zio, "zio_flush: len = %d, eof = %d\n", len, zio_eof (zio));
         if (len > 0) {
             buf = xzmalloc (len + 1);
-            if ((n = zio_fd_read (zio, buf, len + 1)) <= 0) {
+            if ((n = zio_read (zio, buf, len + 1)) <= 0) {
                 if (n < 0) {
                     zio_debug (zio, "zio_read: %s", strerror (errno));
                     rc = -1;
@@ -499,14 +499,17 @@ int zio_flush (zio_t zio)
     return (rc);
 }
 
-
-int zio_read (zio_t zio)
+static int zio_read_cb_common (zio_t zio)
 {
     int n;
-    if ((n = cbuf_write_from_fd (zio->buf, zio->srcfd, -1, NULL)) < 0)
+    if ((n = cbuf_write_from_fd (zio->buf, zio->srcfd, -1, NULL)) < 0) {
+        if (errno == EAGAIN);
+            return (0);
+        zio_debug (zio, "read: %s\n", strerror (errno));
         return (-1);
+    }
 
-    zio_debug (zio, "zio_read: read = %d\n", n);
+    zio_debug (zio, "zio_read_cb: read = %d\n", n);
 
     if (n == 0) {
         zio_set_eof (zio);
@@ -515,15 +518,7 @@ int zio_read (zio_t zio)
 
     zio_flush (zio);
 
-    return (n);
-}
-
-static int zio_read_cb_common (zio_t zio)
-{
-    int rc = zio_read (zio);
-    if ((rc < 0) && (errno == EAGAIN))
-        rc = 0;
-    return (rc);
+    return (0);
 }
 
 static int zio_zloop_read_cb (zloop_t *zl, zmq_pollitem_t *zp, zio_t zio)
@@ -882,7 +877,7 @@ int zio_json_decode (json_object *o, void **pp, bool *eofp)
         else
             *eofp = false;
         if (util_json_object_get_data (o, "data", (uint8_t **) pp, &len) == 0)
-            rc = len; 
+            rc = len;
     }
     return rc;
 }
