+++ src/modules/wreck/Makefile.am
@@ -56,10 +56,11 @@ wrexecd_SOURCES = \
 
 wrexecd_LDADD = \
 	$(top_builddir)/src/bindings/lua/libfluxlua.la \
-	$(top_builddir)/src/common/libzio/libzio.la \
+	$(top_builddir)/src/modules/libzio/libzio.la \
 	$(top_builddir)/src/modules/libmrpc/libmrpc.la \
 	$(top_builddir)/src/common/libflux-internal.la \
 	$(top_builddir)/src/common/libflux-core.la \
+	$(top_builddir)/src/modules/libkz/libkz.la \
 	$(top_builddir)/src/modules/kvs/libkvs.la \
 	$(JSON_LIBS) $(LIBZMQ) $(LIBCZMQ) $(LUA_LIB) $(LIBPTHREAD)
