+++ src/modules/wreck/wrexecd.c
@@ -45,14 +45,14 @@
 #include "src/common/libutil/optparse.h"
 #include "src/common/libutil/jsonutil.h"
 #include "src/common/libutil/xzmalloc.h"
-#include "src/common/libutil/zio.h"
+#include "src/modules/libzio/zio.h"
 #include "src/modules/kvs/kvs.h"
+#include "src/modules/libkz/kz.h"
 
 #include "luastack.h"
 #include "src/bindings/lua/lutil.h"
 #include "src/bindings/lua/kvs-lua.h"
 #include "src/bindings/lua/flux-lua.h"
-#include "src/common/libzio/kz.h"
 
 enum { IN=0, OUT, ERR, NR_IO };
 const char *ionames [] = { "stdin", "stdout", "stderr" };
